requirements.txt
Flask 
SQLAlchemy 
Flask-Migrate 
Flask-WTF 
Flask-Login
request
python-dotenv
